pano_logic_g2: Cleanup and modernize.
